+++ delphi-frontend/src/main/java/au/com/integradev/delphi/type/intrinsic/IntrinsicsInjector.java
@@ -264,7 +264,7 @@ public final class IntrinsicsInjector {
     routine("Sqr").param(type(INTEGER)).returns(type(INTEGER));
     routine("Sqr").param(type(INT64)).returns(type(INT64));
     routine("Str").constParam(TypeFactory.untypedType()).varParam(ANY_STRING);
-    routine("Succ").param(ANY_ORDINAL).returns(type(INTEGER));
+    routine("Succ").param(ANY_ORDINAL).returns(IntrinsicReturnType.argumentByIndex(0));
     routine("Swap").param(type(INTEGER)).returns(type(INTEGER));
     routine("Trunc")
         .param(TypeFactory.untypedType())